moved single data colorbar plots to multiclosure.py/_output.py
